+++ src/components/index.js
@@ -7,7 +7,7 @@ import Footer from "./Footer";
 import PostCard from "./PostCard/PostCard";
 import Button from "./Button";
 import Modal from "./Modal";
-import PostEditor from "./PostEdior";
+import PostEditor from "./PostEditor";
 import PostActions from "./PostActions";
 import CommentsAndShares from "./CommentsAndShares";
 import PeopleListModal from "./PeopleListModal";
